refactor extract skill node (#42)
