+++ frontend/src/features/skillChaining/skills/components/SkillNode.css
@@ -1,0 +1,61 @@
+/* スキルノードの基本スタイル */
+.skill-node {
+  padding: 10px 15px;
+  border-radius: 8px;
+  font-size: 14px;
+  font-weight: 500;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  transition: all 0.2s ease;
+  position: relative;
+  min-width: 120px;
+  text-align: center;
+}
+
+/* ノードのラベル */
+.skill-node-label {
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  max-width: 150px;
+}
+
+/* 選択状態のスタイル */
+.skill-node.selected {
+  box-shadow: 0 0 0 2px #4a90e2;
+}
+
+/* ホバー時のスタイル */
+.skill-node:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.15);
+}
+
+/* ハイライト状態 */
+.skill-node.highlighted-node {
+  z-index: 10;
+  transform: scale(1.05);
+}
+
+/* フェード状態 */
+.skill-node.faded-node {
+  opacity: 0.3;
+}
+
+/* バッジのスタイル（既存のものを拡張） */
+.skill-node .node-badge {
+  position: absolute;
+  top: -8px;
+  right: -8px;
+  background: #ff6b6b;
+  color: white;
+  font-size: 12px;
+  font-weight: bold;
+  min-width: 20px;
+  height: 20px;
+  border-radius: 10px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0 6px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+}
